Add files via upload
Hi guys, I downloaded Google OpenRefine and was playing around with the weather data csv file to try and clean it up a bit. It would be fairly easy to do this manually, but I though i'd learn how to use OpenRefine. I've exported and attached an xlsx file plus the OpenRefine project if you want to have a look. The data is only partially cleaned (details in the project file) and I've made the following changes:
-replaced the "Calm" wind speed data with "0"
-replaced blank wind speed directions with "CALM" where the wind speed was zero
-filled in missing wind speed direction "ENE". Should reasonably be 16 categories for 16 compass points of wind direction. 
-converted columns to numeric and text as necessary (date and time columns not changed)

We should discuss other changes that need to be made, eg index row 189 for the 08/07/2021 has a lot of data missing. Should we fill in this data or just delete the row?