Add files via upload
Hi Guys, I didn't get very far with correlating features in the data set. I had problems loading the csv file that was modified to replace "Calm" with 0 in the wind speed columns. The original csv loads ok though?? Any ideas on this? I'm working in Jupyter Notebooks...
I've uploaded the file anyway if anyone wants a look. I'll keep working on it tonight and tomorrow, but we can progress with the clean data set we already have for now.
Sean