Decision Tree
Basic decision tree code in Jupyter. I had trouble with the Chinese characters in the headers of the original file, so I have modified the csv file for this model, shortening the column names for ease of use. Some of the code had problems converting text to float, so I have also converted the wind direction to numerical, using compass points instead of N, S, E, W etc (used -1 for Calm).
Model is basic but seems to work, accuracy score 0.50.